refactor utils to just be for buttons
